Added Geometry files which will provide geometric functionality.
At this time, only two functions are provided.
coll_ray_tri() returns the point of collision between a ray and a triangle. It assumes that the longest side of the passed triangle ABC is line AB. If this is not guaranteed, the user can call format_tri(A, B, C), and the points will be rearranged so that AB is the longest side without affecting the shape of the triangle.